Lock free (#21)
* Add AtomicIndexPool dependency

* Update nadir

* Use AtomicIndexPool to reduce locks

* Silence global-constructors warning

* Updated to latest AtomicIndexPool

* Support up to 4194303 tasks

* Support up to 4194303 tasks

* Bumped max task count

* Move atomic_index_pool.h from bikeshed.h

* Make local copy of atomic_index_pool.h

* Inlined atmoic index pool

* Store task_id in task

* Lock-free queue, execution order is no longer guaranteed
Loosening the execution order is really not a problem, if you are using
a threadpool to execute the tasks the order is not guaranteed anyway.
If order is needed use dependencies.

* Removed dependency on AtomicIndexPool

* Warning fixes

+++ src/bikeshed.h
@@ -12,7 +12,7 @@ struct SyncPrimitive
 {
     void (*AcquireLock)(SyncPrimitive* primitive);
     void (*ReleaseLock)(SyncPrimitive* primitive);
-    void (*SignalReady)(SyncPrimitive* primitive, uint16_t ready_count);
+    void (*SignalReady)(SyncPrimitive* primitive, uint32_t ready_count);
 };
 
 typedef uint32_t TTaskID;
@@ -26,12 +26,13 @@ enum TaskResult
 
 typedef TaskResult (*TaskFunc)(HShed shed, TTaskID task_id, void* context_data);
 
-uint32_t GetShedSize(uint16_t max_task_count, uint16_t max_dependency_count);
-HShed    CreateShed(void* mem, uint16_t max_task_count, uint16_t max_dependency_count, SyncPrimitive* sync_primitive);
+// Up to 8 388 607 tasks
+uint32_t GetShedSize(uint32_t max_task_count, uint32_t max_dependency_count);
+HShed    CreateShed(void* mem, uint32_t max_task_count, uint32_t max_dependency_count, SyncPrimitive* sync_primitive);
 
-bool CreateTasks(HShed shed, uint16_t task_count, TaskFunc* task_functions, void** task_context_data, TTaskID* out_task_ids);
-void ReadyTasks(HShed shed, uint16_t task_count, const TTaskID* task_ids);
-bool AddTaskDependencies(HShed shed, TTaskID task_id, uint16_t task_count, const TTaskID* dependency_task_ids);
+bool CreateTasks(HShed shed, uint32_t task_count, TaskFunc* task_functions, void** task_context_data, TTaskID* out_task_ids);
+void ReadyTasks(HShed shed, uint32_t task_count, const TTaskID* task_ids);
+bool AddTaskDependencies(HShed shed, TTaskID task_id, uint32_t task_count, const TTaskID* dependency_task_ids);
 
 bool ExecuteOneTask(HShed shed, TTaskID* out_next_ready_task_id);
 void ExecuteAndResolveTask(HShed shed, TTaskID task_id, TTaskID* out_next_ready_task_id);
